Fixed a hint typo: Changed type of beta from int to float

+++ softadapt/algorithms/original_variant.py
@@ -23,7 +23,7 @@ class SoftAdapt(SoftAdaptBase):
 
     """
 
-    def __init__(self, beta: int = 0.1, accuracy_order: int = None):
+    def __init__(self, beta: float = 0.1, accuracy_order: int = None):
         """SoftAdapt class initializer."""
         super().__init__()
         self.beta = beta
